pyproject.toml
[project]
name = "ai-md-editor-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Eno",email = "zeng_eno@qq.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "langchain (>=0.3.15,<0.4.0)",
    "langgraph (>=0.2.65,<0.3.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "motor (>=3.6.1,<4.0.0)",
    "langchain-openai (>=0.3.1,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.2.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false